moved classes to different files; updates #66'
