+++ src/h5cpp/error/stack.cpp
@@ -1,7 +1,7 @@
 //
 // (c) Copyright 2017 DESY,ESS
 //
-// This file is part of h5cpp.
+// This file is part of h5pp.
 //
 // This library is free software; you can redistribute it and/or modify it
 // under the terms of the GNU Lesser General Public License as published
@@ -22,52 +22,28 @@
 // Author: Martin Shetty <martin.shetty@esss.se>
 // Created on: Oct 25, 2017
 //
-#pragma once
 
-#include <h5cpp/core/windows.hpp>
-#include <h5cpp/core/object_handle.hpp>
 #include <h5cpp/error/stack.hpp>
+#include <sstream>
 
 namespace hdf5 {
 namespace error {
 
-class DLL_EXPORT Singleton
+Stack::Stack(std::list<Descriptor> s)
+: std::runtime_error("")
+, contents(s)
 {
- public:
-  static Singleton& instance()
-  {
-    static Singleton singleton_instance;
-    return singleton_instance;
-  }
+  std::stringstream ss;
+  for (auto c : contents)
+    ss << c << "\n";
+  what_message = ss.str();
+}
 
-  void auto_print(bool enable);
-  bool auto_print() const;
 
-  std::string print_stack();
-  Stack extract_stack();
-
-  void throw_exception(const std::string& message);
-
- private:
-  Singleton() {}
-  Singleton(Singleton const&) = delete;
-  void operator=(Singleton const&) = delete;
-
-  bool auto_print_ {true};
-
- private:
-  bool auto_print_enabled() const;
-  void throw_stack();
-  void clear_stack();
-
-  static herr_t to_list(unsigned n,
-                        const H5E_error2_t *err_desc,
-                        std::list<Descriptor>* list);
-};
-
-// prints the explanatory string of an exception. If the exception is nested,
-// recurses to print the explanatory of the exception it holds
-std::string DLL_EXPORT print_exception(const std::exception& e, int level =  0);
+const char* Stack::what() const throw()
+{
+  return what_message.c_str();
+}
 
 
 } // namespace file
