+++ src/h5cpp/error/stack.hpp
@@ -25,49 +25,23 @@
 #pragma once
 
 #include <h5cpp/core/windows.hpp>
-#include <h5cpp/core/object_handle.hpp>
-#include <h5cpp/error/stack.hpp>
+#include <h5cpp/error/descriptor.hpp>
+#include <exception>
+#include <list>
 
 namespace hdf5 {
 namespace error {
 
-class DLL_EXPORT Singleton
+
+class Stack : public std::runtime_error
 {
  public:
-  static Singleton& instance()
-  {
-    static Singleton singleton_instance;
-    return singleton_instance;
-  }
+  Stack(std::list<Descriptor> s);
+  const char* what() const throw();
 
-  void auto_print(bool enable);
-  bool auto_print() const;
-
-  std::string print_stack();
-  Stack extract_stack();
-
-  void throw_exception(const std::string& message);
-
- private:
-  Singleton() {}
-  Singleton(Singleton const&) = delete;
-  void operator=(Singleton const&) = delete;
-
-  bool auto_print_ {true};
-
- private:
-  bool auto_print_enabled() const;
-  void throw_stack();
-  void clear_stack();
-
-  static herr_t to_list(unsigned n,
-                        const H5E_error2_t *err_desc,
-                        std::list<Descriptor>* list);
+  std::list<Descriptor> contents;
+  std::string what_message;
 };
-
-// prints the explanatory string of an exception. If the exception is nested,
-// recurses to print the explanatory of the exception it holds
-std::string DLL_EXPORT print_exception(const std::exception& e, int level =  0);
 
 
 } // namespace file
